% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.correctLogR.R
\name{ascat.correctLogR}
\alias{ascat.correctLogR}
\title{ascat.correctLogR}
\usage{
ascat.correctLogR(ASCATobj, GCcontentfile = NULL, replictimingfile = NULL)
}
\arguments{
\item{ASCATobj}{an ASCAT object}

\item{GCcontentfile}{File containing the GC content around every SNP for increasing window sizes}

\item{replictimingfile}{File containing replication timing at every SNP for various cell lines (optional)}
}
\value{
ASCAT object with corrected tumour logR
}
\description{
Corrects logR of the tumour sample(s) with genomic GC content (replication timing is optional)
}
\details{
Note that probes not present in the GC content file will be lost from the results
}
