% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.R
\name{ascat.GCcorrect}
\alias{ascat.GCcorrect}
\title{ascat.GCcorrect}
\usage{
ascat.GCcorrect(ASCATobj, GCcontentfile = NULL)
}
\arguments{
\item{ASCATobj}{an ASCAT object}

\item{GCcontentfile}{File containing the GC content around every SNP for increasing window sizes}
}
\value{
ASCAT object with corrected tumour logR
}
\description{
Corrects logR of the tumour sample(s) with genomic GC content
}
\details{
Note that probes not present in the GCcontentfile will be lost from the results
}
