% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.plots.R
\name{ascat.plotAdjustedAscatProfile}
\alias{ascat.plotAdjustedAscatProfile}
\title{ascat.plotAdjustedAscatProfile}
\usage{
ascat.plotAdjustedAscatProfile(
  ASCAT_output_object,
  REF,
  y_limit = 5,
  plot_unrounded = FALSE,
  png_prefix = ""
)
}
\arguments{
\item{ASCAT_output_object}{R object generated by the ascat.runAscat function.}

\item{REF}{Can be either "hg19" or "hg38" for standard human genome or a data.frame with three columns: chrom, start and end.}

\item{y_limit}{Optional parameter determining the size of the y axis in the profile (default=5).}

\item{plot_unrounded}{Optional parameter to define whether rounded (default) or unrounded profile (set to TRUE) should be plotted.}

\item{png_prefix}{Optional parameter to add a prefix to png name (can be also used to set a path).}
}
\value{
Plot showing the adjusted (rounded/unrounded) ASCAT profile of the sample
}
\description{
Function plotting the "adjusted" (with realistic chromosome sizes) rounded/unrounded ASCAT profiles over all chromosomes.
}
