% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.plots.R
\name{ascat.plotGenotypes}
\alias{ascat.plotGenotypes}
\title{ascat.plotGenotypes}
\usage{
ascat.plotGenotypes(ASCATobj, title, Tumor_BAF_noNA, Hom, ch_noNA)
}
\arguments{
\item{ASCATobj}{an ASCAT object}

\item{title}{main title of the plot}

\item{Tumor_BAF_noNA}{B-allele frequencies of the tumour sample with removed NA values}

\item{Hom}{Boolean vector denoting homozygous SNPs}

\item{ch_noNA}{vector of probes per chromosome (NA values excluded)}
}
\value{
plot showing classified BAF per sample, with unused SNPs in green, germline homozygous SNPs in blue and all others in red
}
\description{
ascat.plotGenotypes
}
