% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ascat.R
\name{ascat.plotSegmentedData}
\alias{ascat.plotSegmentedData}
\title{ascat.plotSegmentedData}
\usage{
ascat.plotSegmentedData(ASCATobj)
}
\arguments{
\item{ASCATobj}{an ASCAT object (e.g. from ascat.aspcf)}
}
\value{
png files showing raw and segmented tumour logR and BAF
}
\description{
plots the SNP array data before and after segmentation
}

