% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ascat.R
\name{ascat.plotSunrise}
\alias{ascat.plotSunrise}
\title{ascat.plotSunrise}
\usage{
ascat.plotSunrise(d, psi_opt1, rho_opt1)
}
\arguments{
\item{d}{distance matrix for a range of ploidy and tumour percentage values}

\item{psi_opt1}{optimal ploidy}

\item{rho_opt1}{optimal aberrant cell fraction}
}
\value{
plot visualising range of ploidy and tumour percentage values
}
\description{
ascat.plotSunrise
}

