% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.predictGermlineGenotypes.R
\name{ascat.predictGermlineGenotypes}
\alias{ascat.predictGermlineGenotypes}
\title{ascat.predictGermlineGenotypes}
\usage{
ascat.predictGermlineGenotypes(
  ASCATobj,
  platform = "AffySNP6",
  img.dir = ".",
  img.prefix = ""
)
}
\arguments{
\item{ASCATobj}{an ASCAT object}

\item{platform}{used array platform}

\item{img.dir}{directory in which figures will be written}

\item{img.prefix}{prefix for figure names}
}
\value{
predicted germline genotypes
}
\description{
predicts the germline genotypes of samples for which no matched germline sample
is available
}
\details{
Currently possible values for platform:\cr
AffySNP6 (default)\cr
Custom10k\cr
IlluminaASA\cr
IlluminaGSAv3\cr
Illumina109k\cr
IlluminaCytoSNP\cr
IlluminaCytoSNP850k\cr
Illumina610k\cr
Illumina660k\cr
Illumina700k\cr
Illumina1M\cr
Illumina2.5M\cr
IlluminaOmni5\cr
IlluminaGDACyto-8\cr
Affy10k\cr
Affy100k\cr
Affy250k_sty\cr
Affy250k_nsp\cr
AffyOncoScan\cr
AffyCytoScanHD\cr
HumanCNV370quad\cr
HumanCore12\cr
HumanCoreExome24\cr
HumanOmniExpress12\cr
IlluminaOmniExpressExome\cr
}
