% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.prepareHTS.R
\name{ascat.synchroniseFiles}
\alias{ascat.synchroniseFiles}
\title{Synchronise SNPs across files}
\usage{
ascat.synchroniseFiles(
  samplename,
  tumourLogR_file,
  tumourBAF_file,
  normalLogR_file,
  normalBAF_file
)
}
\arguments{
\item{samplename}{String, name of the sample.}

\item{tumourLogR_file}{File where LogR from the tumour will be read and overwritten.}

\item{tumourBAF_file}{File where BAF from the tumour will be read and overwritten.}

\item{normalLogR_file}{File where LogR from the normal will be read and overwritten.}

\item{normalBAF_file}{File where BAF from the normal will be read and overwritten.}
}
\description{
Synchronise SNPs across files
}
\author{
tl
}
