% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.runAscat.R
\name{make_segments}
\alias{make_segments}
\title{make_segments}
\usage{
make_segments(r, b)
}
\arguments{
\item{r}{segmented logR}

\item{b}{segmented BAF}
}
\value{
segments of constant logR and BAF including their lengths
}
\description{
Function to make segments of constant LRR and BAF.\cr
This function is more general and does not depend on specific ASPCF output, it can also handle segmention performed on LRR and BAF separately
}
\keyword{internal}
