% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.correctLogR.R
\name{ascat.GCcorrect}
\alias{ascat.GCcorrect}
\title{ascat.GCcorrect}
\usage{
ascat.GCcorrect(ASCATobj, GCcontentfile = NULL)
}
\arguments{
\item{ASCATobj}{an ASCAT object}

\item{GCcontentfile}{File containing the GC content around every SNP for increasing window sizes}
}
\description{
Function kept for backward compatibility, please use ascat.correctLogR instead
}
