% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.prepareHTS.R
\name{ascat.getBAFsAndLogRs}
\alias{ascat.getBAFsAndLogRs}
\title{Obtain BAF and LogR from the allele counts.}
\usage{
ascat.getBAFsAndLogRs(
  samplename,
  tumourAlleleCountsFile.prefix,
  normalAlleleCountsFile.prefix,
  tumourLogR_file,
  tumourBAF_file,
  normalLogR_file,
  normalBAF_file,
  g1000file.prefix,
  gender,
  chrom_names = c(1:22, "X"),
  minCounts = 8,
  BED_file = NA,
  probloci_file = NA,
  seed = as.integer(Sys.time())
)
}
\arguments{
\item{samplename}{String, name of the sample.}

\item{tumourAlleleCountsFile.prefix}{Prefix of the allele counts files for the tumour.}

\item{normalAlleleCountsFile.prefix}{Prefix of the allele counts files for the normal.}

\item{tumourLogR_file}{File where LogR from the tumour will be written.}

\item{tumourBAF_file}{File where BAF from the tumour will be written.}

\item{normalLogR_file}{File where LogR from the normal will be written.}

\item{normalBAF_file}{File where BAF from the normal will be written.}

\item{g1000file.prefix}{Prefix to where 1000 Genomes reference files can be found.}

\item{gender}{Gender information, either 'XX' (=female) or 'XY' (=male).}

\item{chrom_names}{A vector with allowed chromosome names (optional, default=c(1:22,'X')).}

\item{minCounts}{Minimum depth, in mormal, required for a SNP to be considered (optional, default=8).}

\item{BED_file}{A BED file for only looking at SNPs within specific intervals (optional, default=NA).}

\item{probloci_file}{A file (chromosome <tab> position; no header) containing specific loci to ignore (optional, default=NA).}

\item{seed}{A seed to be set for when randomising the alleles (optional, default=as.integer(Sys.time())).}
}
\description{
Obtain BAF and LogR from the allele counts.
}
\author{
dw9, sd11, tl
}
