% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.plots.R
\name{ascat.plotSegmentedData}
\alias{ascat.plotSegmentedData}
\title{ascat.plotSegmentedData}
\usage{
ascat.plotSegmentedData(
  ASCATobj,
  img.dir = ".",
  img.prefix = "",
  logr.y_values = c(-2, 2)
)
}
\arguments{
\item{ASCATobj}{an ASCAT object (e.g. from ascat.aspcf)}

\item{img.dir}{directory in which figures will be written}

\item{img.prefix}{prefix for figure names}

\item{logr.y_values}{define Y min and max values for logR track (optional; default: c(-2, 2))}
}
\value{
png files showing raw and segmented tumour logR and BAF
}
\description{
plots the SNP array data before and after segmentation
}
